+++ CHANGELOG.md
@@ -23,6 +23,7 @@
   * Added option ``"nosandbox"`` for browser factory
   * Added option ``"sendSyncDefaultTimeout"`` for browser factory
   * Added option ``"ignoreCertificateErrors"`` for browser factory
+  * Added option ``"customFlags"`` for browser factory
 * Bug fixes:
   * Fixed user agent string for browser factory